Move LanguageSelector out of the LanguageSelectionDialog. Also, re-write the test case.
