--- osu.Game.Rulesets.Karaoke.Tests/Graphics/TestSceneLanguageSelectionDialog.cs
@@ -1,61 +1,0 @@
-﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
-// See the LICENCE file in the repository root for full licence text.
-
-using System.Globalization;
-using NUnit.Framework;
-using osu.Framework.Allocation;
-using osu.Framework.Bindables;
-using osu.Framework.Graphics;
-using osu.Framework.Graphics.Containers;
-using osu.Framework.Testing;
-using osu.Game.Overlays;
-using osu.Game.Rulesets.Karaoke.Beatmaps;
-using osu.Game.Rulesets.Karaoke.Graphics.UserInterface;
-using osu.Game.Rulesets.Karaoke.Tests.Beatmaps;
-using osu.Game.Screens.Edit;
-using osu.Game.Tests.Visual;
-
-namespace osu.Game.Rulesets.Karaoke.Tests.Graphics
-{
-    public class TestSceneLanguageSelectionDialog : OsuManualInputManagerTestScene
-    {
-        protected override Container<Drawable> Content { get; } = new Container { RelativeSizeAxes = Axes.Both };
-
-        private DialogOverlay dialogOverlay;
-
-        private LanguageSelectionDialog dialog;
-
-        [BackgroundDependencyLoader]
-        private void load()
-        {
-            var beatmap = new TestKaraokeBeatmap(Ruleset.Value);
-            var karaokeBeatmap = new KaraokeBeatmapConverter(beatmap, new KaraokeRuleset()).Convert();
-            var editorBeatmap = new EditorBeatmap(karaokeBeatmap);
-            Dependencies.Cache(editorBeatmap);
-
-            base.Content.AddRange(new Drawable[]
-            {
-                Content,
-                dialogOverlay = new DialogOverlay()
-            });
-
-            Dependencies.Cache(dialogOverlay);
-        }
-
-        [SetUp]
-        public void SetUp() => Schedule(() =>
-        {
-            var language = new Bindable<CultureInfo>(new CultureInfo("ja"));
-            Child = dialog = new LanguageSelectionDialog
-            {
-                Current = language,
-            };
-        });
-
-        [SetUpSteps]
-        public void SetUpSteps()
-        {
-            AddStep("show dialog", () => dialog.Show());
-        }
-    }
-}
